requirements.txt
jira>=1.0.0,<2.0.0
PyGithub>=1.0.0,<2.0.0
